Add files via upload
This file defines the routes for user management using Flask's Blueprint feature. It contains the following endpoints:
users.py:
GET: Retrieves a list of all users.
GET: Retrieves a specific user by their ID. If the user is not found, it returns a 404 error.
POST: Creates a new user with a unique ID, name, email, and age. The new user data is extracted from the request's JSON payload.
PUT: Updates the details of an existing user based on their ID. If the user is not found, it returns a 404 error.
DELETE: Deletes a user by their ID. If the user is not found, it returns a 404 error.
The data is stored in a list named users, and each user is represented as a dictionary containing fields like id, name, email, and age.
app.py:
Registers the Blueprint: It imports the users_bp Blueprint from the  file and registers.
Error handling: It handles 404 errors by returning a JSON response with a message indicating that the resource was not found.
Runs the application: The app is configured to run on port 3000 when executed directly.
3. auth.py
This file defines an authentication function that checks if the request contains a valid token. The functionality includes:

authenticate_token(): Checks for the presence of an Authorization header in the incoming request. If the header is missing or if the token value doesn't match the expected token (Visca Barca), it returns a JSON response indicating that the request is unauthorized with a 401 status code.
The token-based authentication is a simple mechanism to secure endpoints by verifying the validity of requests.